data/requirements.txt
tensorflow==2.18.1
pillow
numpy